requirements.txt
aiohttp==3.9.1
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.3
click==8.1.7
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.9.0
dnspython==2.5.0
frozenlist==1.4.1
fsspec==2023.12.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.16
langsmith==0.0.83
llama-index==0.9.39
marshmallow==3.20.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
openai==1.10.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.25
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
voyageai==0.1.7
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
